+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.5] Q3 2022
+ - renamed from ICS-DeviceManagement to omnect github orga
+
 ## [0.8.4] Q3 2022
  - add get_sdk_version_string() API
  - bump to azure-iot-sdk-sys 0.5.2